Back-End/requirements.txt
fastapi
uvicorn
python-multipart
supabase
pydantic
requests
numpy
python-dotenv
instructor
groq
google-genai
PyJWT